third_party/blink/web_tests/platform/generic/fast/events/inputText-never-fired-on-keydown-cancel-expected.txt

Test that ensure textInput events should not fired when preceding key events are cancelled

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS window.targetNode is receivedEventTarget
PASS null is receivedEventTarget
PASS window.targetNode is receivedEventTarget
PASS null is receivedEventTarget
PASS window.targetNode is receivedEventTarget
PASS null is receivedEventTarget
PASS successfullyParsed is true

TEST COMPLETE
